+++ src/Model/PaymentList.php
@@ -1,0 +1,34 @@
+<?php
+
+declare(strict_types=1);
+
+namespace amcintosh\FreshBooks\Model;
+
+use Spatie\DataTransferObject\Attributes\CastWith;
+use Spatie\DataTransferObject\Attributes\MapFrom;
+use Spatie\DataTransferObject\Casters\ArrayCaster;
+use Spatie\DataTransferObject\DataTransferObject;
+use amcintosh\FreshBooks\Model\Payment;
+
+/**
+ * Results of Payments list call containing list of payments and pagination data.
+ *
+ * @package amcintosh\FreshBooks\Model
+ * @link https://www.freshbooks.com/api/payments
+ */
+class PaymentList extends DataTransferObject
+{
+    public const RESPONSE_FIELD = 'payments';
+
+    public int $page;
+
+    public int $pages;
+
+    #[MapFrom('per_page')]
+    public int $perPage;
+
+    public int $total;
+
+    #[CastWith(ArrayCaster::class, itemType: Payment::class)]
+    public array $payments;
+}